+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 3.46.0
+  - feat: Adds business verification result notification endpoint
+
+# 3.45.0
+  - feat: Add Romanian language support
+
+# 3.44.0
+  - feat: Enhance InternalProjectViewSet to support retrieval of project language details
+
 # 3.43.0
   - feat: Adds cached plan-status endpoint for retail consumption